+++ UNO/main.c
@@ -1,28 +1,74 @@
 /*
- * Slave.c
+ * final_project_slave
+ * main.c 
+ * 
+ *		BL40A1812 Introduction to Embedded Systems
+ *		final project - slave
  *
- * Created: 17/04/2024
- * Author : Group_14
- * Used examples: Exercise 10, Exercise 7
- * used pins: Buzzer (GNR, 9 "PB1" --> Buzzer) USART = (10, 11, 12, 13, GNR --> Mega)
+ * Created: 4/2024
+ * Author : Group 14
  */ 
 
-/* ATmega328p "UNO" */
-
+/* DEFINES COME FROM Ex. 9 */
 #define F_CPU 16000000UL
 #define FOSC 16000000UL // Clock Speed
 #define BAUD 9600
 #define MYUBRR (FOSC/16/BAUD-1)
 
+#include <string.h>
 #include <avr/io.h>
 #include <util/delay.h>
 #include <util/setbaud.h>
 #include <stdio.h>
 #include <avr/interrupt.h>
-#include <stdlib.h> // turha?
-#include <string.h>
+#include <util/delay.h>
 
-static void USART_init(uint16_t ubrr) // unsigned int
+/* STATE MACHINE VALUES */
+#define BUZZER_OFF 1
+#define BUZZER_ON 2
+
+#define STRING_LENGTH 20
+
+/* functions for the timer/counter1 compare match A interrupt vector */
+ISR
+(TIMER1_COMPA_vect)
+{
+    TCNT1 = 0; // reset timer counter
+}
+
+int caltop(int f, int n) // Counts top - page 128
+{
+    long f_cpu = F_CPU;
+    return ((f_cpu)/(2*n*f));
+}
+
+/*
+static void buzzerSong()
+{
+    TCCR1A |= (1 << 6);
+    // D3
+    OCR1A = caltop(100, 1);
+    _delay_ms(10000);
+    OCR1A = caltop(147, 1);
+    _delay_ms(10000);
+    printf("First delay");
+    _delay_ms(1000);
+    printf("Second delay");
+    
+    // C3 SHARP
+    printf("C3 sharp");
+    OCR1A = caltop(139, 1);
+
+    _delay_ms(1000);
+    // C3
+    OCR1A = caltop(130, 1);
+    _delay_ms(1000);
+    _delay_ms(10000);
+}
+*/
+
+static void
+USART_init(uint16_t ubrr) // unsigned int
 {
     /* Set baud rate in the USART Baud Rate Registers (UBRR) */
     UBRR0H = (unsigned char) (ubrr >> 8);
@@ -36,25 +82,27 @@ static void USART_init(uint16_t ubrr) // unsigned int
     
 }
 
-/* functions for the timer/counter1 compare match A interrupt vector */
-ISR (TIMER1_COMPA_vect)
-{
-    TCNT1 = 0; // reset timer counter
-}
-
-static void USART_Transmit(unsigned char data, FILE *stream)
+static void
+USART_Transmit(unsigned char data, FILE *stream)
 {
     /* Wait until the transmit buffer is empty*/
-    while(!(UCSR0A & (1 << UDRE0))) {;}
+    while(!(UCSR0A & (1 << UDRE0)))
+    {
+        ;
+    }
     
     /* Put the data into a buffer, then send/transmit the data */
     UDR0 = data;
 }
 
-static char USART_Receive(FILE *stream)
+static char
+USART_Receive(FILE *stream)
 {
-    /* Wait until the transmit buffer is empty - Wait for data to be received */
-    while(!(UCSR0A & (1 << UDRE0))) {;}
+    /* Wait until the transmit buffer is empty*/
+    while(!(UCSR0A & (1 << UDRE0)))
+    {
+        ;
+    }
     
     /* Get the received data from the buffer */
     return UDR0;
@@ -64,11 +112,31 @@ static char USART_Receive(FILE *stream)
 FILE uart_output = FDEV_SETUP_STREAM(USART_Transmit, NULL, _FDEV_SETUP_WRITE);
 FILE uart_input = FDEV_SETUP_STREAM(NULL, USART_Receive, _FDEV_SETUP_READ);
 
-int main(void)
+void receiveMessageFromMaster(char *string) {
+	char spi_receive_data[STRING_LENGTH];
+	for(int8_t spi_data_index = 0; spi_data_index < STRING_LENGTH; spi_data_index++)
+	{
+		while(!(SPSR & (1 << SPIF)))
+		{
+			/* wait until the transmission is complete */
+			;
+		}
+		spi_receive_data[spi_data_index] = SPDR; // receive byte from the SPI data register
+
+	}
+	strcpy(string, spi_receive_data);
+}
+
+int 
+main(void)
 {
+
+    int state = 1; 
+    /* set MISO as output, pin 12 (PB4)*/
+    DDRB  = (1 << PB4);
+    /* set SPI enable */
+    SPCR  = (1 << 6);
     
-    /*-----------Buzzer--------------*/
-    /* set up the ports and pins */
     DDRB |= (1 << PB1); // OC1A is located in digital pin 9
     
     // Enable interrupts
@@ -78,67 +146,58 @@ int main(void)
     TCCR1B  = 0; // reset timer/counter 1
     TCNT1   = 0;
     TCCR1A |= (1 << 6); // set compare output mode to toggle
-    // TCCR1A |= 0b01000000;
-    // TCCR1A |= 0x40;
-    
+        
     // mode 9 phase correct
     TCCR1A |= (1 << 0); // set register A WGM[1:0] bits
-    // TCCR1A |= 0b00000001;
     TCCR1B |= (1 << 4); // set register B WBM[3:2] bits
-    // TCCR1B |= 0b00010000;
-    
+        
     TIMSK1 |= (1 << 1); // enable compare match A interrupt
-    // TIMSK1 |= 0b00000100;
-    
+        
     OCR1A = 15297; // C5 523 Hz, no prescaler
-    //OCR1A = 2462;   // A7 3250 Hz, no prescaler, calculated
-    //OCR1A = 2440;   // A7 3250 Hz, no prescaler, empirical
-    //OCR1A = 1016;   // B2  123 Hz, 64 prescaler
+    TCCR1B |= (1 << 0); // set prescaling to 1 (no prescaling)
     
-    /*---------------------------------*/
-    
-    
-    /* set MISO as output, pin 12 (PB4)*/
-    DDRB  = (1 << PB4);
-    /* set SPI enable */
-    SPCR  = (1 << 6);
+    TCCR1A &= ~(1 << 6);
     
     // initialize the UART with 9600 BAUD
     USART_init(MYUBRR);
-    
+        
     // redirect the stdin and stdout to UART functions
     stdout = &uart_output;
     stdin = &uart_input;
+	
+    char spi_receive_data[STRING_LENGTH];
     
-    char spi_send_data[20] = "slave to master\n";
-    char spi_receive_data[20];
-    
-    /* send message to master and receive message from master */
-    while (1)
-    {
-        
-        for(int8_t spi_data_index = 0; spi_data_index < sizeof(spi_send_data); spi_data_index++)
-        {
-            
-            SPDR = spi_send_data[spi_data_index]; // send byte using SPI data register
-            
-            while(!(SPSR & (1 << SPIF)))
-            {
-                /* wait until the transmission is complete */
-                ;
-            }
-            spi_receive_data[spi_data_index] = SPDR; // receive byte from the SPI data register
-
+    while (1) 
+    { 
+		receiveMessageFromMaster(spi_receive_data);
+        switch(state) {
+            case BUZZER_OFF:
+                 _delay_ms(1000);
+				 // If message received is ON, set buzzer on
+                  if (strstr(spi_receive_data,("ON")))
+                  {
+                      // ATmega datasheet page 131 Table 16.1
+                      printf("Buzzer turning on\n");
+                      TCCR1A |= (1 << 6);
+                      OCR1A = caltop(100, 1);
+                       state = BUZZER_ON;
+                  }
+                  break;
+            case BUZZER_ON:
+			// If message received is OFF, set buzzer off
+             if (strstr(spi_receive_data,("OFF")))
+                   {
+                       // ATmega datasheet page 131 Table 16.1 
+                       TCCR1A &= ~(1 << 6);
+                       printf("Buzzer turning off\n");
+                       state = BUZZER_OFF;
+                   }
+                   break;
+			default:
+				printf("ERROR\n");
+				break;
         }
-        
-        printf(spi_receive_data);
-        if (!strcmp(spi_receive_data,("1")))
-        {
-            TCCR1B |= (1 << 0); // set prescaling to 1 (no prescaling)
-        }else{
-            TCCR1B |= (0 << 0); // Buzzer off???
-        }
-    }
-    
+    }      
     return 0;
 }
+
